+++ CHANGELOG.md
@@ -32,3 +32,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed 
 
 - Event System moved to BBUnity Events
+
+## [0.3.0] - 2020-11-03
+### Removed 
+
+- Removed the ability for CallbackHandler to find Global Interfaces 
+
+### Added 
+
+- Added protection if statement on the Process
+
+### Edited
+
+- Changed Call to Process on the CallbackHandler